// src/utils/enums.ts
export enum navbarPaths {
  Home = "",
  About = "about",
  Calendar = "calendar",
  Notes = "notes",
  Results = "results",
  Sale = "sale",
  Videos = "videos",
}

export enum DeviceType {
  SmallDesktop = "SmallDesktop",
  MediumDesktop = "MediumDesktop",
  LargeDesktop = "LargeDesktop",
  Tablet = "Tablet",
  Mobile = "Mobile",
}